Moved all 2.#.# files into src folder as finished
