--- test/README.md
@@ -1,2 +1,0 @@
-<h1>Code here is currently being worked on but is not released.</h1>
-<body>Don't use any code here on your worlds! It is broken!</body>
